requirements.txt
# Core Django
Django>=5.0.4,<6.0
python-decouple>=3.8

# Image Processing
Pillow>=10.0.0
reportlab>=4.0.0

# Production Server (optional)
gunicorn>=21.0.0

# Development Tools (optional)
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.3.0
